Refactor the server for better testability.
Made the server struct unexported.
Moved the creation of the server into a helper method.
Added port as a property of the server struct.
Refactored the serving process to start on a different thread and
return the server variable for ease of shutting down the server. The
main thread now listens on SIGINT signals and shuts down the server
gracefully when it receives one.